Function to print the last half of a string

+++ 0x05-pointers_arrays_strings/7-puts_half.c
@@ -5,7 +5,7 @@
  */
 void puts_half(char *str)
 {
-int i, half;
+long i, half;
 i = 0;
 while (str[i] != '\0')
 {
